refactor(architecture): decouple model architecture from lightning module
- Rename `src/mlcast/models/convgru.py` to `base.py`.
- Refactor `RadarLightningModel` into a generic `NowcastLightningModule` that accepts an injected PyTorch `nn.Module`.
- Extract the core `EncoderDecoder` logic into a separate file (`src/mlcast/modules/convgru_modules.py`) and rename it to `ConvGruModel`.
- Update `src/mlcast/configs.py` to use Fiddle to inject `ConvGruModel` into `NowcastLightningModule`.

This completes Phase 2A of the restructuring plan, establishing a clean separation of concerns between the training orchestrator and the underlying neural network architecture.
